split the lib.rs into several files
the file was becoming too long
